+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 - Do not mutate `pattern` keywords in response schema validation. :issue:`2749`
 - Generate empty header values during the coverage phase.
 - Generate negative values for `minItems` & `maxItems` during the coverage phase.
+- Use `default` value as valid input during the coverage phase.
 
 .. _v3.39.14: